fix a bug in a GdkPixbuf method
The notify function for `gdk_pixbuf_new_from_data` has its arguments flipped with respect to the assumed order for a closure, which resulted in a segfault when we try to free the GdkPixbuf.

Note that due to a different bug, we don't actually free GdkPixbufs right now (along with many other things). This is one part of a fix for that.