optionally initialize a GtkCanvas immediately
This is helpful for precompilation. GtkCanvas was one of the few widgets where being realized was important to calling its functions. With this we can avoid popping up windows during precompilation.

This also changes the behavior of the width and height arguments of the GtkCanvas constructor. We now set the "content area" size rather than the initial size of the widget. I think this is more useful because it sets a minimum size for the widget. The initial size can still be set by calling "size_request" on the widget after construction.