In callbacks, convert inputs and outputs between pointers and objects (#45)
In callbacks we have had to convert pointer arguments to GObjects and return pointers rather than the GObjects. This modifies the GI-generated callback functions to do that automatically, which makes the user-supplied callbacks shorter and cleaner. This also automatically adds a reference to GObject pointers on the output where appropriate.

Also adds some crude documentation for GtkListView, GtkColumnView, GtkCustomFilter, and GtkCustomSorter and miscellaneous other fixes.