support GtkSizeGroup and add tests (#53)
Add push! and delete! methods for GtkSizeGroup and tests for a method that returns a GList with transfer=none.

Also put a limit on the version of libpng, which prevents tests from failing on Windows x86. I assume this is a bug in one of the libraries and will look into it further later.