Version 0.7 (#74)
* require julia 1.10

* minor stuff

* improve test coverage (#70)

* Use new dialogs (breaking) (#49)

In the newer versions of GTK v4, new dialogs were introduced and the old ones were deprecated, including GtkDialog and GtkMessageDialog. There isn't much difference in user-facing behavior, but since the old ones were deprecated it's less likely bugs will be fixed. This PR migrates to the new dialogs and cleans up our API a little.

I left the file dialogs alone after looking in the C code and realizing that the new dialogs are based on GtkFileChooserNative, like the old code.

* GtkTextView: use pointer or Ref instead of _GtkTextIter struct (breaking) (#62)

* GtkTextView: use pointer or ref instead of _GtkTextIter struct

This is more consistent with everything else in the package

* random stuff

* support adding widgets to gutter

* fix GtkTextIter auto-generated methods, better support methods for all non-opaque structs

* a bit of tidying

* remove unneeded version check

* add simple TextView example

* update auto-generated code for Adwaita

* simplify generation script for Adwaita

* update auto-generated code

* regenerate code