In `gobject_move_ref`, replace weak reference with strong reference in `gc_preserve_glib` (#89)
This prevents these objects, which are "subclasses" such as GtkCanvas, from being garbage
collected at all. Previously, the code was supposed to intercept the garbage collection
and do this "weak to strong" conversion if the object survives being unreferenced by
Julia, but https://github.com/JuliaGtk/Gtk4.jl/issues/87 seems to indicate a problem.
This is a band-aid fix for that issue.