use libgirepository in Glib_jll >=2.80 (#107)
In GLib 2.80, libgirepository was moved in from gobject-introspection and rewritten. This adapts GI.jl to use it.

For some reason, if the generation scripts are run sequentially in the same Julia session, as was done in "gen_all.jl", libgirepository can't reliably figure out what shared library is associated with a given namespace. I assume this can be fixed here but for now we will have to run the generation script for each namespace in a separate Julia session.